Move sptbinq into tqueue
